Send different message based on current time
